+++ Packages/com.verasl.water-system/Scripts/Data/WaterSettingsData.cs
@@ -14,7 +14,7 @@ namespace WaterSystem.Data
 		public GeometryType waterGeomType; // The type of geometry, either vertex offset or tessellation
         public ReflectionType refType = ReflectionType.PlanarReflection; // How the reflecitons are generated
 		// planar
-		public PlanerReflections.PlanarReflectionSettings planarSettings; // Planar reflection settings
+		public PlanarReflections.PlanarReflectionSettings planarSettings; // Planar reflection settings
 		// cubemap
 		public Cubemap cubemapRefType; // custom cubemap reference
