+++ Packages/com.verasl.water-system/Scripts/Rendering/PlanarReflections.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3971d89a730cd5846a1944ab0228bf76
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
