--- Packages/com.verasl.water-system/Scripts/Rendering/PlanerReflections.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 62d2ec625e93e8d40b72db412b9db648
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
