+++ Packages/manifest.json
@@ -1,20 +1,17 @@
 {
   "registry": "https://staging-packages.unity.com",
   "dependencies": {
-    "com.unity.ads": "2.0.8",
-    "com.unity.analytics": "3.2.2",
-    "com.unity.burst": "0.2.4-preview.40",
+    "com.unity.analytics": "3.3.2",
+    "com.unity.burst": "0.2.4-preview.45",
     "com.unity.cinemachine": "2.3.3",
-    "com.unity.collab-proxy": "1.2.15",
-    "com.unity.mathematics": "0.0.12-preview.19",
+    "com.unity.mathematics": "0.0.12-preview.21",
     "com.unity.package-manager-ui": "2.1.1",
-    "com.unity.postprocessing": "2.1.2",
-    "com.unity.render-pipelines.core": "file:/Users/andremcgrail/Documents/UnityProjects/Github/SRP_Master/com.unity.render-pipelines.core",
-    "com.unity.render-pipelines.lightweight": "file:/Users/andremcgrail/Documents/UnityProjects/Github/SRP_Master/com.unity.render-pipelines.lightweight",
-    "com.unity.shadergraph": "file:/Users/andremcgrail/Documents/UnityProjects/Github/SRP_Master/com.unity.shadergraph",
+    "com.unity.postprocessing": "2.1.3",
+    "com.unity.render-pipelines.core": "file:../../SRP/com.unity.render-pipelines.core",
+    "com.unity.render-pipelines.lightweight": "file:../../SRP/com.unity.render-pipelines.lightweight",
+    "com.unity.shadergraph": "file:../../SRP/com.unity.shadergraph",
     "com.unity.textmeshpro": "1.3.0",
     "com.unity.timeline": "1.0.0",
-    "com.unity.xr.legacyinputhelpers": "1.0.0",
     "com.unity.modules.ai": "1.0.0",
     "com.unity.modules.animation": "1.0.0",
     "com.unity.modules.assetbundle": "1.0.0",
