+++ ProjectSettings/BurstAotSettings_Android.json
@@ -1,0 +1,11 @@
+{
+    "MonoBehaviour": {
+        "m_Enabled": true,
+        "m_EditorHideFlags": 0,
+        "m_Name": "",
+        "m_EditorClassIdentifier": "",
+        "DisableOptimisations": false,
+        "DisableSafetyChecks": true,
+        "DisableBurstCompilation": false
+    }
+}
