+++ ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2019.1.0b1
-m_EditorVersionWithRevision: 2019.1.0b1 (83b3ba1f99df)
+m_EditorVersion: 2019.2.0a6
+m_EditorVersionWithRevision: 2019.2.0a6 (5612105fae18)
